portfolio_app/requirements.txt
streamlit==1.28.2
plotly==5.18.0
Pillow==10.1.0
pyyaml==6.0.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0
librosa==0.10.1
--extra-index-url https://download.pytorch.org/whl/cpu
